404: Embeded Google Webmaster Tools custom search form.
Feature request at:
http://p2pu.lighthouseapp.com/projects/71002-lernanta/tickets/608

The domain for the form gets set to localhost, even though it is
set to be "http://p2pu.org". This is probably due to the fact
that development server run at localhost.

Need to check if the domain is properly set when running on a
public server.